backend-mln/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0
pydantic==2.4.0
